requirements.txt
Django==2.2.7
python-slugify==4.0.0
pytz==2019.3
sqlparse==0.3.0
text-unidecode==1.3
